got delete and edit to show only if it was created by the logged in user

+++ DevPrep/Controllers/ConceptsController.cs
@@ -46,6 +46,7 @@ namespace DevPrep.Controllers
                     ConceptId = c.Id,
                     ConceptName = c.Name,
                     SoftwareLanguageId = id,
+                    ApplicationUserId = c.ApplicationUserId,
                     UsefulLinks = c.UsefulLinks.ToList(),
                     Descriptions = c.Descriptions.ToList()
                 }).ToListAsync();
@@ -194,8 +195,16 @@ namespace DevPrep.Controllers
             try
             {
                 var usefulLinks = await _context.UsefulLinks.Where(ul => ul.ConceptId == id).ToListAsync();
+                var descriptions = await _context.Descriptions.Where(d => d.ConceptId == id).ToListAsync();
                 var softwareLangId = concept.SoftwareLanguageId;
-                _context.UsefulLinks.RemoveRange(usefulLinks);
+                if (usefulLinks.Count > 0)
+                {
+                    _context.UsefulLinks.RemoveRange(usefulLinks);
+                }
+                if (descriptions.Count > 0)
+                {
+                    _context.Descriptions.RemoveRange(descriptions);
+                }
